+++ .env.example
@@ -77,7 +77,6 @@ DS_ENABLE_CAUSE_INTERESTS_PAGE=
 DS_ENABLE_DYNAMIC_CAMPAIGNS_PAGE=
 DS_ENABLE_HIDE_CAMPAIGNS=
 DS_ENABLE_QUIZ_RESULT_PAGE=
-DS_ENABLE_NEW_HOMEPAGE=
 DS_ENABLE_NPS_SURVEY=
 DS_ENABLE_REFER_FRIENDS_INCENTIVE=
 DS_ENABLE_SITEWIDE_NPS_SURVEY=
