+++ config/feature-flags.php
@@ -17,7 +17,6 @@ return [
     'cause_preferences' => env('DS_ENABLE_CAUSE_INTERESTS_PAGE', false),
     'dynamic_explore_campaigns' => env('DS_ENABLE_DYNAMIC_CAMPAIGNS_PAGE', false),
     'hide_campaigns' => env('DS_ENABLE_HIDE_CAMPAIGNS', false),
-    'new_homepage' => env('DS_ENABLE_NEW_HOMEPAGE', false),
     'nps_survey' => env('DS_ENABLE_NPS_SURVEY', false),
     'quiz_result_page' =>  env('DS_ENABLE_QUIZ_RESULT_PAGE', false),
     'refer_friends_incentive' => env('DS_ENABLE_REFER_FRIENDS_INCENTIVE', false),
