+++ resources/assets/components/pages/HomePage/HomePage.js
@@ -1,96 +1,445 @@
-/* global window */
-
-import React from 'react';
 import { get } from 'lodash';
+import gql from 'graphql-tag';
+import PropTypes from 'prop-types';
+import { css } from '@emotion/core';
+import { React, Fragment } from 'react';
 
+import PageQuery from '../PageQuery';
 import sponsorList from './sponsor-list';
 import Modal from '../../utilities/Modal/Modal';
+import * as NewsletterImages from './NewsletterImages';
+import HomePageArticleGallery from './HomePageArticleGallery';
 import SiteFooter from '../../utilities/SiteFooter/SiteFooter';
-import { contentfulImageUrl, featureFlag } from '../../../helpers';
+import HomePageCampaignGallery from './HomePageCampaignGallery';
+import PrimaryButton from '../../utilities/Button/PrimaryButton';
+import { pageCardFragment } from '../../utilities/PageCard/PageCard';
 import TypeFormEmbed from '../../utilities/TypeFormEmbed/TypeFormEmbed';
 import DelayedElement from '../../utilities/DelayedElement/DelayedElement';
+import { campaignCardFragment } from '../../utilities/CampaignCard/CampaignCard';
 import SiteNavigationContainer from '../../SiteNavigation/SiteNavigationContainer';
+import AnalyticsWaypoint from '../../utilities/AnalyticsWaypoint/AnalyticsWaypoint';
 import DismissableElement from '../../utilities/DismissableElement/DismissableElement';
 import TrafficDistribution from '../../utilities/TrafficDistribution/TrafficDistribution';
+import { campaignCardFeaturedFragment } from '../../utilities/CampaignCard/CampaignCardFeatured';
+import {
+  contentfulImageUrl,
+  featureFlag,
+  isAuthenticated,
+  tailwind,
+} from '../../../helpers';
 
-import './home-page.scss';
-
-class HomePage extends React.Component {
-  constructor() {
-    super();
-    this.state = get(window.HOMEPAGE, 'fields', {});
+const HOME_PAGE_QUERY = gql`
+  query HomePageQuery($preview: Boolean!) {
+    page: homePage(preview: $preview) {
+      id
+      title
+      subTitle
+      coverImage {
+        url
+      }
+      campaigns {
+        ...CampaignCard
+        ...CampaignCardFeatured
+      }
+      articles {
+        ...PageCard
+      }
+      additionalContent
+    }
   }
 
-  renderGalleryBlock = (block, index) => {
-    const { id, type, staffPick } = block;
+  ${campaignCardFragment}
+  ${campaignCardFeaturedFragment}
+  ${pageCardFragment}
+`;
 
-    const isCampaign = type === 'campaign';
+/**
+ * Impact statistic component for homepage header banner.
+ *
+ * @param {Object}
+ */
+const ImpactStatistic = ({ campaignName, impactLabel, impactValue }) => (
+  <div className="px-4 py-5 text-center">
+    <p className="font-league-gothic leading-none mb-1 text-5xl text-teal-300 uppercase">
+      {impactValue}
+    </p>
+    <p className="font-bold m-0 mb-2 text-lg text-white uppercase">
+      {impactLabel}
+    </p>
+    <p className="italic m-0 text-white">{campaignName}</p>
+  </div>
+);
 
-    const fields = isCampaign ? block : block.fields;
-    const { slug, title, coverImage } = fields;
+ImpactStatistic.propTypes = {
+  campaignName: PropTypes.string.isRequired,
+  impactLabel: PropTypes.string.isRequired,
+  impactValue: PropTypes.string.isRequired,
+};
 
-    // Set more generous crop dimensions for 'featured' (first) homepage tile.
-    const imageCrop = index === 0 ? '1600' : '800';
+/**
+ * Newsletter item for newsletter section.
+ *
+ * @param {Object}
+ */
+const NewsletterItem = ({ content, image, link, title }) => (
+  <div className="flex flex-col h-full">
+    <a className="block mb-4 mx-auto no-underline" href={link.url}>
+      <img
+        alt={`${title} newsletter badge`}
+        css={css`
+          height: 180px;
+          width: 180px;
+        `}
+        src={image}
+      />
+    </a>
 
-    return (
-      <article className="tile" key={id}>
-        <a
-          className="wrapper"
-          href={`/us/${isCampaign ? 'campaigns/' : ''}${slug}`}
-        >
-          {staffPick ? (
-            <div className="tile__flag -staff-pick">STAFF PICK</div>
+    <h3 className="mb-2">
+      <a className="text-white hover:text-yellow-300" href={link.url}>
+        {title}
+      </a>
+    </h3>
+
+    <p className="mb-4 flex-grow text-white">{content}</p>
+
+    <a
+      className="font-normal text-white hover:text-yellow-300 underline hover:no-underline"
+      data-label={`newsletter_cta_${title.toLowerCase()}`}
+      href={link.url}
+      rel="noopener noreferrer"
+      target="_blank"
+    >
+      {link.copy}
+    </a>
+  </div>
+);
+
+NewsletterItem.propTypes = {
+  content: PropTypes.string.isRequired,
+  image: PropTypes.string.isRequired,
+  link: PropTypes.shape({
+    copy: PropTypes.string.isRequired,
+    url: PropTypes.string.isRequired,
+  }).isRequired,
+  title: PropTypes.string.isRequired,
+};
+
+/**
+ * Home Page template.
+ *
+ * @param {Object}
+ */
+const HomePageTemplate = ({ articles, campaigns, coverImage, title }) => {
+  const tailwindGray = tailwind('colors.gray');
+  const tailwindScreens = tailwind('screens');
+
+  const centerHorizontalRule = css`
+    @media (min-width: ${tailwindScreens.md}) {
+      margin-top: -2px;
+      top: 50%;
+    }
+  `;
+
+  const headerBackgroundStyles = coverImage
+    ? css`
+        background-image: url(${contentfulImageUrl(
+          coverImage.url,
+          '400',
+          '775',
+          'fill',
+        )});
+
+        @media (min-width: ${tailwindScreens.md}) {
+          background-image: url(${contentfulImageUrl(
+            coverImage.url,
+            '700',
+            '700',
+            'fill',
+          )});
+        }
+
+        @media (min-width: ${tailwindScreens.lg}) {
+          background-image: url(${contentfulImageUrl(
+            coverImage.url,
+            '1440',
+            '539',
+            'fill',
+          )});
+        }
+      `
+    : null;
+
+  return (
+    <Fragment>
+      {/* @TODO: Once EmotionJS supports shorthand syntax for React.Fragment, switch <Fragment> out for <> syntax! */}
+      <SiteNavigationContainer />
+
+      <main>
+        <article data-test="home-page">
+          <header role="banner" className="bg-white">
+            <AnalyticsWaypoint name="impact_section_top" />
+
+            <div
+              className="base-12-grid bg-gray-200 pt-3 md:pt-6"
+              css={css`
+                background-position: center center;
+                background-repeat: no-repeat;
+                background-size: cover;
+                padding-bottom: 250px;
+
+                @media (min-width: ${tailwindScreens.md}) {
+                  padding-bottom: 100px;
+                }
+
+                ${headerBackgroundStyles}
+              `}
+            >
+              <h1
+                className="font-league-gothic font-normal grid-wide m-0 px-2 md:px-0 lg:px-16 py-24 text-yellow-500 text-center uppercase"
+                css={css`
+                  font-size: 84px;
+                  line-height: 1.1;
+
+                  @media (min-width: ${tailwindScreens.md}) {
+                    font-size: 120px;
+                  }
+                `}
+              >
+                {title}
+              </h1>
+            </div>
+
+            <div
+              className="base-12-grid"
+              css={css`
+                margin-top: -250px;
+
+                @media (min-width: ${tailwindScreens.md}) {
+                  margin-top: -100px;
+                }
+              `}
+            >
+              <div className="bg-blurple-500 py-4 grid-wide xl:grid xl:grid-cols-3">
+                <ImpactStatistic
+                  campaignName="Teens for Jeans"
+                  impactLabel="Jeans Donated"
+                  impactValue="5 million"
+                />
+
+                <ImpactStatistic
+                  campaignName="GTFO (Get The Filter Out)"
+                  impactLabel="Cigarette Butts Collected"
+                  impactValue="3.7 million"
+                />
+
+                <ImpactStatistic
+                  campaignName="Power to the Period"
+                  impactLabel="Period Products Donated"
+                  impactValue="585,965"
+                />
+              </div>
+            </div>
+
+            <AnalyticsWaypoint name="impact_section_bottom" />
+          </header>
+
+          {campaigns ? (
+            <section
+              className="base-12-grid bg-gray-100 py-8"
+              css={css`
+                background: linear-gradient(
+                  to bottom,
+                  rgba(255, 255, 255, 1) 25%,
+                  ${tailwindGray['100']}
+                );
+              `}
+              data-test="campaigns-section"
+            >
+              <div className="grid-wide text-center">
+                <AnalyticsWaypoint name="campaign_section_top" />
+
+                <h2 className="mb-6 relative">
+                  <span className="bg-white font-league-gothic font-normal leading-tight inline-block px-6 relative text-3xl md:text-4xl uppercase z-10">
+                    Take Action!
+                  </span>
+                  <span
+                    className="absolute bg-purple-500 block h-1 w-full z-0"
+                    css={centerHorizontalRule}
+                  />
+                </h2>
+
+                <p className="mb-3 text-lg">
+                  Choose a campaign below to make an impact and enter for a
+                  chance to{' '}
+                  <a
+                    href="/us/about/easy-scholarships"
+                    className="font-normal text-blurple-500 hover:text-blurple-300 underline hover:no-underline"
+                    data-label="campaign_section_earn_scholarships"
+                  >
+                    earn scholarships
+                  </a>
+                  .
+                </p>
+
+                <p className="mb-6 lg:mb-8 mt-0 text-lg">
+                  (Talk about a win-win.)
+                </p>
+
+                {/*
+                // @TODO: After volunteer credits launch use this copy for campaigns section:
+                <p className="mb-3 text-lg">
+                  Choose a campaign below to make an impact{' '}
+                  <a
+                    href="/us/about/easy-scholarships"
+                    className="font-normal text-blurple-500 hover:text-blurple-300 underline hover:no-underline"
+                    data-label="campaign_section_earn_scholarships"
+                  >
+                    win scholarships
+                  </a>
+                  , and earn volunteer credits for school.
+                </p>
+
+                <p className="mb-6 lg:mb-8 mt-0 text-lg">
+                  Talk about a win-win-win.
+                </p>
+                */}
+
+                <HomePageCampaignGallery campaigns={campaigns} />
+
+                <PrimaryButton
+                  attributes={{ 'data-label': 'campaign_section_show_more' }}
+                  className="mt-8 py-4 px-8 text-lg"
+                  href="/us/campaigns"
+                  text="See More Campaigns"
+                />
+
+                <AnalyticsWaypoint name="campaign_section_bottom" />
+              </div>
+            </section>
           ) : null}
 
-          <div className="tile__meta">
-            <h1 className="tile__title">{title}</h1>
-            <p className="tile__tagline">{fields.tagline || fields.subTitle}</p>
-          </div>
+          <article
+            className="base-12-grid bg-purple-700 py-8 lg:py-12"
+            data-test="newsletters-cta"
+          >
+            <div className="grid-wide text-center">
+              <AnalyticsWaypoint name="newsletter_cta_top" />
 
-          <img
-            src={contentfulImageUrl(
-              coverImage.url,
-              imageCrop,
-              imageCrop,
-              'fill',
-            )}
-            alt={coverImage.description || `${title} cover image`}
-          />
-        </a>
-      </article>
-    );
-  };
+              <h2 className="font-league-gothic font-normal mb-4 tracking-wide text-4xl text-white uppercase">
+                Impact &amp; Inspiration For Your Inbox
+              </h2>
 
-  render() {
-    const { blocks, title, subTitle } = this.state;
+              <p className="text-lg text-white">
+                Subscribe to a DoSomething newsletter.
+              </p>
 
-    if (!blocks) {
-      return null;
-    }
+              <ul className="md:col-gap-8 lg:col-gap-16 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 mt-8 row-gap-8">
+                <li className="text-white">
+                  <NewsletterItem
+                    content="How DoSomething members are changing the world, and how you can too."
+                    title="Community"
+                    image={NewsletterImages.Community}
+                    link={{
+                      copy: 'Join Us',
+                      url:
+                        'https://join.dosomething.org/#page-block-71pdxfteosc',
+                    }}
+                  />
+                </li>
 
-    return (
-      <>
-        <SiteNavigationContainer />
+                <li className="text-white">
+                  <NewsletterItem
+                    content="Earn the chance to win scholarships for volunteering. No essay or GPA required...ever."
+                    title="Scholarships"
+                    image={NewsletterImages.Scholarships}
+                    link={{
+                      copy: 'Find Scholarships',
+                      url:
+                        'https://join.dosomething.org/#page-block-sy2dstz4b6',
+                    }}
+                  />
+                </li>
 
-        <main>
-          <article className="home-page">
-            <header role="banner" className="header header--home">
-              <div className="wrapper">
-                <h1 className="header__title">{title}</h1>
-                <h2 className="header__subtitle">{subTitle}</h2>
+                <li className="text-white">
+                  <NewsletterItem
+                    content="Read the headlines. Change the headlines."
+                    title="News"
+                    image={NewsletterImages.News}
+                    link={{
+                      copy: 'Get News',
+                      url:
+                        'https://join.dosomething.org/#page-block-ay8syiyhro5',
+                    }}
+                  />
+                </li>
+
+                <li className="text-white">
+                  <NewsletterItem
+                    content="Boost your mood. Boost your knowledge. Boost yourself into action!"
+                    title="Lifestyle"
+                    image={NewsletterImages.Lifestyle}
+                    link={{
+                      copy: 'Get Inspired',
+                      url:
+                        'https://join.dosomething.org/#page-block-85qyo7ii8fh',
+                    }}
+                  />
+                </li>
+              </ul>
+
+              <AnalyticsWaypoint name="newsletter_cta_bottom" />
+            </div>
+          </article>
+
+          {articles ? (
+            <section
+              className="base-12-grid bg-gray-100 py-8"
+              data-test="articles-section"
+            >
+              <div className="grid-wide text-center">
+                <AnalyticsWaypoint name="article_section_top" />
+
+                <h2 className="mb-6 relative">
+                  <span className="bg-gray-100 font-league-gothic font-normal leading-tight inline-block px-6 relative text-3xl md:text-4xl uppercase z-10">
+                    Read About It
+                  </span>
+                  <span
+                    className="absolute bg-purple-500 block h-1 w-full z-0"
+                    css={centerHorizontalRule}
+                  />
+                </h2>
+
+                <HomePageArticleGallery articles={articles} />
+
+                <PrimaryButton
+                  attributes={{ 'data-label': 'article_section_show_more' }}
+                  className="mt-8 py-4 px-8 text-lg"
+                  href="https://lets.dosomething.org/"
+                  text="See More Articles"
+                />
+
+                <AnalyticsWaypoint name="article_section_bottom" />
               </div>
-            </header>
-
-            <section className="home-page__gallery">
-              {blocks.map(this.renderGalleryBlock)}
             </section>
+          ) : null}
 
-            <section className="container--sponsors md:w-3/4 mx-auto px-3 py-8">
-              <h4>Sponsors</h4>
+          {/* @TODO: Need to remove the top/bottom padding from base-12-grid class and
+          let components add their own padding otherwise it is hard to override. */}
+          <section
+            className="base-12-grid bg-white py-8"
+            data-test="sponsors-section"
+          >
+            <div className="grid-wide text-center">
+              <AnalyticsWaypoint name="sponsor_section_top" />
+
+              <h2 className="font-bold mb-3 text-base text-center text-gray-500 uppercase">
+                Sponsors
+              </h2>
               <ul>
                 {sponsorList.map(sponsor => (
-                  <li key={sponsor.name}>
+                  <li key={sponsor.name} className="inline-block mx-6 my-3">
                     <img
+                      className="opacity-25"
                       src={contentfulImageUrl(sponsor.image, '125', '40')}
                       title={sponsor.name}
                       alt={sponsor.name}
@@ -98,36 +447,92 @@ class HomePage extends React.Component {
                   </li>
                 ))}
               </ul>
-            </section>
-          </article>
 
-          {featureFlag('nps_survey') ? (
-            <TrafficDistribution percentage={5} feature="homepage_nps_survey">
-              <DismissableElement
-                name="nps_survey"
-                render={(handleClose, handleComplete) => (
-                  <DelayedElement delay={5}>
-                    <Modal onClose={handleClose} trackingId="SURVEY_MODAL">
-                      <TypeFormEmbed
-                        displayType="modal"
-                        typeformUrl="https://dosomething.typeform.com/to/iEdy7C"
-                        queryParameters={{
-                          northstar_id: get(window.AUTH, 'id', null),
-                        }}
-                        onSubmit={handleComplete}
-                      />
-                    </Modal>
-                  </DelayedElement>
-                )}
-              />
-            </TrafficDistribution>
-          ) : null}
-        </main>
+              <AnalyticsWaypoint name="sponsor_section_bottom" />
+            </div>
+          </section>
 
-        <SiteFooter />
-      </>
-    );
-  }
-}
+          {isAuthenticated() ? null : (
+            <article
+              className="base-12-grid bg-yellow-500 py-16"
+              data-test="signup-cta"
+            >
+              <div className="xl:flex grid-wide xl:items-center text-center">
+                <AnalyticsWaypoint name="join_cta_top" />
+
+                <div className="text-left xl:w-8/12">
+                  <h1 className="font-bold text-2xl">
+                    Join our youth-led movement for good
+                  </h1>
+                  <p className="text-lg">
+                    Make an impact with millions of young people, and earn a
+                    chance to win scholarships.
+                  </p>
+                </div>
+
+                <div className="flex-grow">
+                  <PrimaryButton
+                    attributes={{ 'data-label': 'signup_cta_authorize' }}
+                    className="mt-8 xl:m-0 py-4 px-16 text-lg xl:ml-auto"
+                    href="/authorize"
+                    text="Join Now"
+                  />
+                </div>
+
+                <AnalyticsWaypoint name="join_cta_bottom" />
+              </div>
+            </article>
+          )}
+        </article>
+
+        {featureFlag('nps_survey') ? (
+          <TrafficDistribution percentage={5} feature="homepage_nps_survey">
+            <DismissableElement
+              name="nps_survey"
+              render={(handleClose, handleComplete) => (
+                <DelayedElement delay={5}>
+                  <Modal onClose={handleClose} trackingId="SURVEY_MODAL">
+                    <TypeFormEmbed
+                      displayType="modal"
+                      typeformUrl="https://dosomething.typeform.com/to/iEdy7C"
+                      queryParameters={{
+                        northstar_id: get(window.AUTH, 'id', null),
+                      }}
+                      onSubmit={handleComplete}
+                    />
+                  </Modal>
+                </DelayedElement>
+              )}
+            />
+          </TrafficDistribution>
+        ) : null}
+      </main>
+
+      <SiteFooter />
+    </Fragment>
+  );
+};
+
+HomePageTemplate.propTypes = {
+  articles: PropTypes.arrayOf(PropTypes.object),
+  campaigns: PropTypes.arrayOf(PropTypes.object),
+  coverImage: PropTypes.shape({
+    url: PropTypes.string.isRequired,
+  }),
+  title: PropTypes.string,
+};
+
+HomePageTemplate.defaultProps = {
+  articles: null,
+  campaigns: null,
+  coverImage: null,
+  title: 'We Are A Youth-Led Movement For Good',
+};
+
+const HomePage = () => (
+  <PageQuery query={HOME_PAGE_QUERY}>
+    {page => <HomePageTemplate {...page} />}
+  </PageQuery>
+);
 
 export default HomePage;
