--- resources/assets/components/pages/HomePage/home-page.scss
@@ -1,61 +1,0 @@
-@import '../../../scss/next-toolbox';
-
-.home-page {
-  .home-page__gallery {
-    display: grid;
-
-    @include media($medium) {
-      grid-template-columns: repeat(4, 1fr);
-
-      .tile:first-of-type {
-        grid-column: span 2;
-        grid-row: span 2;
-      }
-    }
-  }
-
-  .header {
-    padding-top: 72px;
-    padding-bottom: 50px;
-  }
-
-  .header--home {
-    .header__title {
-      text-align: center;
-    }
-
-    .header__subtitle {
-      font-size: $font-medium;
-      font-weight: $weight-normal;
-      text-align: center;
-    }
-  }
-
-  .container--sponsors {
-    text-align: center;
-
-    h4 {
-      color: theme('colors.gray.200');
-    }
-
-    ul {
-      list-style: none;
-      padding: 0;
-    }
-
-    li {
-      display: inline-block;
-      margin: theme('spacing.3') theme('spacing.6');
-    }
-
-    img {
-      vertical-align: middle;
-      opacity: 0.2;
-
-      @include media($mobile) {
-        max-width: 90px;
-        max-height: 30px;
-      }
-    }
-  }
-}
